chore: move layout metadata schema map to components
